+++ lib/workload/stateful/stacks/shared/constructs/event-source/index.ts
